Breaking renderer formats apart
